Moves files around in prep for refactor
