+++ package.json
@@ -10,7 +10,7 @@
     "@mui/lab": "^5.0.0-alpha.120",
     "@mui/material": "^5.11.10",
     "@mui/styles": "^5.11.9",
-    "@mui/x-data-grid": "^4.0.1",
+    "@mui/x-data-grid": "^6.0.0",
     "@types/jest": "^29.4.0",
     "@types/node": "^18.14.0",
     "@types/react": "^18.0.28",
