+++ src/containers/RecipeEdit.js
@@ -1,7 +1,7 @@
 import React from "react";
-import { withRouter } from "react-router-dom";
+import {withRouter} from "react-router-dom";
 import useIngredientLO from "data/useIngredientLO";
-import RecipeEdit from "../views/cook/RecipeEdit";
+import RecipeEdit from "../features/RecipeEdit/RecipeEdit";
 
 export default withRouter(({match}) => {
     const id = parseInt(match.params.id, 10);
