+++ src/features/RecipeEdit/RecipeFormController.tsx
@@ -1,0 +1,6 @@
+import * as React from "react"
+import RecipeEdit from "./RecipeEdit";
+
+export const RecipeFormController = () => {
+    return (<RecipeEdit />)
+}
